Changed the PI controller for a PD controller
Changed the data in the csv files of the data directory, all the
columns of the PI controller were deleted and replaced with five
new tests using a Pd controller.
This decission was taken due to the results obtained in the field
trial, in this trial the results with the PI controller were not
as expected.